Moving common 0D and 1D RO CONFIG to _MembraneBase
