+++ watertap/unit_models/membrane_base.py
@@ -1,0 +1,140 @@
+###############################################################################
+# WaterTAP Copyright (c) 2021, The Regents of the University of California,
+# through Lawrence Berkeley National Laboratory, Oak Ridge National
+# Laboratory, National Renewable Energy Laboratory, and National Energy
+# Technology Laboratory (subject to receipt of any required approvals from
+# the U.S. Dept. of Energy). All rights reserved.
+#
+# Please see the files COPYRIGHT.md and LICENSE.md for full copyright and license
+# information, respectively. These files are also available online at the URL
+# "https://github.com/watertap-org/watertap/"
+#
+###############################################################################
+
+from enum import Enum, auto
+from pyomo.common.config import ConfigBlock, ConfigValue, In
+from idaes.core import UnitModelBlockData, useDefault, MaterialBalanceType,\
+        EnergyBalanceType, MomentumBalanceType
+from idaes.core.util.config import is_physical_parameter_block
+import idaes.logger as idaeslog
+
+_log = idaeslog.getLogger(__name__)
+
+class ConcentrationPolarizationType(Enum):
+    none = auto()                    # simplified assumption: no concentration polarization
+    fixed = auto()                   # simplified assumption: concentration polarization modulus is a user specified value
+    calculated = auto()              # calculate concentration polarization (concentration at membrane interface)
+
+
+class MassTransferCoefficient(Enum):
+    none = auto()                    # mass transfer coefficient not utilized for concentration polarization effect
+    fixed = auto()                   # mass transfer coefficient is a user specified value
+    calculated = auto()              # mass transfer coefficient is calculated
+    # TODO: add option for users to define their own relationship?
+
+
+class PressureChangeType(Enum):
+    fixed_per_stage = auto()         # pressure drop across membrane channel is a user-specified value
+    fixed_per_unit_length = auto()   # pressure drop per unit length across membrane channel is a user-specified value
+    calculated = auto()              # pressure drop across membrane channel is calculated
+
+class _MembraneBaseData(UnitModelBlockData):
+    """
+    Membrane-based filtration base class
+    """
+    CONFIG = UnitModelBlockData.CONFIG()
+
+    CONFIG.declare("property_package", ConfigValue(
+        default=useDefault,
+        domain=is_physical_parameter_block,
+        description="Property package to use for control volume",
+        doc="""Property parameter object used to define property calculations,
+    **default** - useDefault.
+    **Valid values:** {
+    **useDefault** - use default package from parent model or flowsheet,
+    **PhysicalParameterObject** - a PhysicalParameterBlock object.}"""))
+
+    CONFIG.declare("property_package_args", ConfigBlock(
+        implicit=True,
+        description="Arguments to use for constructing property packages",
+        doc="""A ConfigBlock with arguments to be passed to a property block(s)
+    and used when constructing these.
+    **default** - None.
+    **Valid values:** {
+    see property package for documentation.}"""))
+
+    CONFIG.declare("material_balance_type", ConfigValue(
+            default=MaterialBalanceType.useDefault,
+            domain=In(MaterialBalanceType),
+            description="Material balance construction flag",
+            doc="""Indicates what type of mass balance should be constructed,
+    **default** - MaterialBalanceType.useDefault.
+    **Valid values:** {
+    **MaterialBalanceType.useDefault - refer to property package for default
+    balance type
+    **MaterialBalanceType.none** - exclude material balances,
+    **MaterialBalanceType.componentPhase** - use phase component balances,
+    **MaterialBalanceType.componentTotal** - use total component balances,
+    **MaterialBalanceType.elementTotal** - use total element balances,
+    **MaterialBalanceType.total** - use total material balance.}"""))
+
+    CONFIG.declare("energy_balance_type", ConfigValue(
+        default=EnergyBalanceType.useDefault,
+        domain=In(EnergyBalanceType),
+        description="Energy balance construction flag",
+        doc="""Indicates what type of energy balance should be constructed.
+    **default** - EnergyBalanceType.useDefault.
+    **Valid values:** {
+    **EnergyBalanceType.useDefault - refer to property package for default
+    balance type
+    **EnergyBalanceType.none** - exclude energy balances,
+    **EnergyBalanceType.enthalpyTotal** - single enthalpy balance for material,
+    **EnergyBalanceType.enthalpyPhase** - enthalpy balances for each phase,
+    **EnergyBalanceType.energyTotal** - single energy balance for material,
+    **EnergyBalanceType.energyPhase** - energy balances for each phase.}"""))
+
+    CONFIG.declare("momentum_balance_type", ConfigValue(
+            default=MomentumBalanceType.pressureTotal,
+            domain=In(MomentumBalanceType),
+            description="Momentum balance construction flag",
+            doc="""Indicates what type of momentum balance should be constructed,
+    **default** - MomentumBalanceType.pressureTotal.
+    **Valid values:** {
+    **MomentumBalanceType.none** - exclude momentum balances,
+    **MomentumBalanceType.pressureTotal** - single pressure balance for material,
+    **MomentumBalanceType.pressurePhase** - pressure balances for each phase,
+    **MomentumBalanceType.momentumTotal** - single momentum balance for material,
+    **MomentumBalanceType.momentumPhase** - momentum balances for each phase.}"""))
+
+    CONFIG.declare("has_pressure_change", ConfigValue(
+            default=False,
+            domain=In([True, False]),
+            description="Pressure change term construction flag",
+            doc="""Indicates whether terms for pressure change should be
+    constructed,
+    **default** - False.
+    **Valid values:** {
+    **True** - include pressure change terms,
+    **False** - exclude pressure change terms.}"""))
+
+    CONFIG.declare("pressure_change_type", ConfigValue(
+        default=PressureChangeType.fixed_per_stage,
+        domain=In(PressureChangeType),
+        description="Pressure change term construction flag",
+        doc="""
+        Indicates what type of pressure change calculation will be made. To use any of the 
+        ``pressure_change_type`` options to account for pressure drop, the configuration keyword 
+        ``has_pressure_change`` must also be set to ``True``. Also, if a value is specified for pressure 
+        change, it should be negative to represent pressure drop. 
+        
+        **default** - ``PressureChangeType.fixed_per_stage`` 
+
+
+    .. csv-table::
+        :header: "Configuration Options", "Description"
+
+        "``PressureChangeType.fixed_per_stage``", "Specify an estimated value for pressure drop across the membrane feed channel"
+        "``PressureChangeType.fixed_per_unit_length``", "Specify an estimated value for pressure drop per unit length across the membrane feed channel"
+        "``PressureChangeType.calculated``", "Allow model to perform calculation of pressure drop across the membrane feed channel"
+
+"""))
